requirements.txt
pip~=25.2
attrs~=25.3.0
discord.py[voice]
discord.py
aiohttp
python-dotenv
youtube-search-python
yt-dlp
requests
